backend/pyproject.toml
[project]
name = "fireshire-backend"
version = "0.1.0"
description = "Fire-resilient landscaping zone visualizer — backend API"
authors = [
    {name = "Robert Head"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi>=0.135.2,<0.136.0",
    "uvicorn[standard]>=0.42.0,<0.43.0",
    "pydantic-settings>=2.13.1,<3.0.0",
    "httpx>=0.28.1,<0.29.0",
    "anthropic>=0.52.0,<1.0.0",
    "sqlalchemy[asyncio]>=2.0.48,<3.0.0",
    "asyncpg>=0.31.0,<0.32.0",
    "greenlet>=3.3.2,<4.0.0",
]


[tool.poetry]
packages = [{include = "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "httpx>=0.28.1,<0.29.0",
    "respx>=0.22.0,<0.23.0",
]
local = [
    "alembic>=1.18.4,<2.0.0",
]
